python-app/requirements.txt
numpy
scipy
colour
psutil
python-rtmidi
zerorpc
python-socketio
pyserial
